pyproject.toml
[tool.poetry]
name = "asyncpygame"
version = "0.1.0.dev0"
description = "Async library for PyGame"
authors = ["Nattōsai Mitō <flow4re2c@gmail.com>"]
license = "MIT"
readme = 'README.md'
repository = 'https://github.com/asyncgui/asyncpygame'
homepage = 'https://github.com/asyncgui/asyncpygame'
keywords = ['async', 'pygame']
classifiers=[
    'Development Status :: 2 - Pre-Alpha',
    'License :: OSI Approved :: MIT License',
    'Intended Audience :: Developers',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Topic :: Software Development :: Libraries',
    'Operating System :: OS Independent',
]
packages = [
    { include = "asyncpygame", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
asyncgui = "~0.6"
pygame-ce = "^2.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
flake8 = "^6.1.0"

[tool.poetry.group.doc.dependencies]
sphinx = "^7.2.6"
sphinx-autobuild = "^2021.3.14"
furo = "^2023.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
xfail_strict = true
addopts = "--maxfail=4 --strict-markers"
